cloud functions split from a single file to multiple files
